Improvements in the code-
- I made the following corrections:

a) In the .mint-form input selector, the font-size property is missing a unit. It should be font-size: 17px;
b) Removed the comment (/* border: solid 1px red; */) as it was not needed.
c) Adjusted the indentation for better readability.
d) Added relevant comments to the code for better user understanding and knowlede.